Adding new secured routes to instructor side bar to the correct pages
